Overall refactoring for state system. Implemented calm state

+++ src/com/oom/game/main/entities/state/SleepingState.java
@@ -84,7 +84,7 @@ public class SleepingState implements NPCState {
     }
 
     @Override
-    public int getSpeed() {
+    public double getSpeed() {
         return 0;
     }
 }
